pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-orchestrator"
version = "1.0.0"
description = "Multi-agent AI orchestrator for IT diagnostics and engineering workflows"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "MentalVibez", email = ""}
]
keywords = ["ai", "agents", "orchestrator", "llm", "workflow", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "boto3>=1.29.0",
    "openai>=1.3.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
]

[project.urls]
Homepage = "https://github.com/MentalVibez/ai-agent-orchestrator"
Documentation = "https://github.com/MentalVibez/ai-agent-orchestrator#readme"
Repository = "https://github.com/MentalVibez/ai-agent-orchestrator"
Issues = "https://github.com/MentalVibez/ai-agent-orchestrator/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_ignores = false
ignore_missing_imports = true
follow_imports = "skip"

